+++ crates/api_models/src/gsm.rs
@@ -1,3 +1,4 @@
+use common_enums::{ErrorCategory, ErrorSubCategory};
 use utoipa::ToSchema;
 
 use crate::enums::Connector;
@@ -30,63 +31,6 @@ pub struct GsmCreateRequest {
     pub error_category: Option<ErrorCategory>,
     /// sub-category in which error belongs to
     pub error_sub_category: Option<ErrorSubCategory>,
-}
-
-#[derive(
-    Clone,
-    Copy,
-    Debug,
-    strum::Display,
-    PartialEq,
-    Eq,
-    serde::Serialize,
-    serde::Deserialize,
-    strum::EnumString,
-    ToSchema,
-)]
-#[serde(rename_all = "snake_case")]
-#[strum(serialize_all = "snake_case")]
-pub enum ErrorCategory {
-    IssuerDecline,
-    ProcessorDecline,
-    TechnicalErrorWithProcessor,
-}
-
-#[derive(
-    Clone,
-    Copy,
-    Debug,
-    strum::Display,
-    PartialEq,
-    Eq,
-    serde::Serialize,
-    serde::Deserialize,
-    strum::EnumString,
-    ToSchema,
-)]
-#[serde(rename_all = "snake_case")]
-#[strum(serialize_all = "snake_case")]
-pub enum ErrorSubCategory {
-    AuthFailure,
-    CardDecline,
-    ExpiredCard,
-    Frm,
-    General,
-    IncorrectAccount,
-    IncorrectCard,
-    IncorrectCvv,
-    IncorrectData,
-    InsufficientFunds,
-    Partial,
-    Restricted,
-    BadRequest,
-    ConfigurationError,
-    CustomerError,
-    IncorrectAuth,
-    Unauthorized,
-    ServerError,
-    Timeout,
-    Upstream,
 }
 
 #[derive(Debug, serde::Deserialize, serde::Serialize, ToSchema)]
@@ -207,7 +151,7 @@ pub struct GsmResponse {
     /// error message unified across the connectors
     pub unified_message: Option<String>,
     /// category in which error belongs to
-    pub error_category: Option<String>,
+    pub error_category: Option<ErrorCategory>,
     /// sub-category in which error belongs to
-    pub error_sub_category: Option<String>,
+    pub error_sub_category: Option<ErrorSubCategory>,
 }
