+++ crates/openapi/src/openapi.rs
@@ -317,8 +317,6 @@ Never share your secret api keys. Keep them guarded and secure.
         api_models::gsm::GsmDeleteResponse,
         api_models::gsm::GsmResponse,
         api_models::gsm::GsmDecision,
-        api_models::gsm::ErrorCategory,
-        api_models::gsm::ErrorSubCategory,
         api_models::payments::AddressDetails,
         api_models::payments::BankDebitData,
         api_models::payments::AliPayQr,
@@ -606,6 +604,8 @@ Never share your secret api keys. Keep them guarded and secure.
         api_models::blocklist::ToggleBlocklistResponse,
         api_models::blocklist::ListBlocklistQuery,
         api_models::enums::BlocklistDataKind,
+        api_models::enums::ErrorCategory,
+        api_models::enums::ErrorSubCategory,
         api_models::webhook_events::EventListItemResponse,
         api_models::webhook_events::EventRetrieveResponse,
         api_models::webhook_events::OutgoingWebhookRequestContent,
